# faster & type stable `invperm` and `isperm` for tuples shorter that 16 elements