--- a/base/combinatorics.jl
+++ b/base/combinatorics.jl
@@ -244,12 +244,12 @@ function invperm(p::Union{Tuple{},Tuple{Int},Tuple{Int,Int}})
 end
 
 function invperm(P::NTuple{N,T}) where {N,T}
-    isperm(p) || throw(ArgumentError("argument is not a permutation"))
+    isperm(P) || throw(ArgumentError("argument is not a permutation"))
     ntuple(Val(N)) do i
         for j in P
             @inbounds P[j]==i && return j
         end
-        return zero(T) #
+        return first(P) # DO NOT REMOVE: necessary for type stable code_gen
     end
 end