--- a/base/combinatorics.jl
+++ b/base/combinatorics.jl
@@ -249,7 +249,6 @@ function invperm(P::NTuple{N,T}) where {N,T}
             @inbounds P[j]==i && return j
         end
         throw(ArgumentError("argument is not a permutation"))
-        return first(P) # DO NOT REMOVE: necessary for type stable code_gen
     end
 end