#### Regarding speed
small simple benchmarking code
```julia
using Random
using BenchmarkTools

function fast_invperm(P::NTuple{N,T}) where {N,T}
    isperm(P) || throw(ArgumentError("argument is not a permutation"))
    ntuple(Val(N)) do i
        for j in P
            @inbounds P[j]==i && return j
        end
        return first(P) # DO NOT REMOVE: necessary for type stable code_gen
    end
end

function time_impl()
    for N in 3:10:50
        T=(shuffle(1:N)...,)
        @show @benchmark invperm($T)
        @show @benchmark fast_invperm($T)
    end
end

time_impl()
```

results:

```
#= ....108 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(253.927 ns)
#= ....109 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(41.746 ns)
#= ....108 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(450.132 ns)
#= ....109 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(90.729 ns)
#= ....108 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(637.290 ns)
#= ....109 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(194.659 ns)
#= ....108 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(1.296 μs)
#= ....109 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(303.951 ns)
#= ....108 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(1.612 μs)
#= ....109 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(475.883 ns)
```


#### Regarding type stability
```julia
@code_warntype invperm((1,3,2))
Variables
  #self#::Core.Compiler.Const(invperm, false)
  a::Tuple{Int64,Int64,Int64}

Body::Tuple{Vararg{Int64,N} where N}
1 ─ %1 = Core._apply_iterate(Base.iterate, Base.vect, a)::Array{Int64,1}
│   %2 = Base.invperm(%1)::Array{Int64,1}
│   %3 = Core._apply_iterate(Base.iterate, Core.tuple, %2)::Tuple{Vararg{Int64,N} where N}
└──      return %3
```


```julia
@code_warntype fast_invperm((1,3,2))

Variables
  #self#::Core.Compiler.Const(fast_invperm, false)
  P::Tuple{Int64,Int64,Int64}
  #9::var"#9#10"{Tuple{Int64,Int64,Int64}}

Body::Tuple{Int64,Int64,Int64}
1 ─       Core.NewvarNode(:(#9))
│   %2  = Main.isperm(P)::Bool
└──       goto #3 if not %2
2 ─       goto #4
3 ─ %5  = Main.ArgumentError("argument is not a permutation")::ArgumentError
└──       Main.throw(%5)
4 ┄ %7  = Main.:(var"#9#10")::Core.Compiler.Const(var"#9#10", false)
│   %8  = Core.typeof(P)::Core.Compiler.Const(Tuple{Int64,Int64,Int64}, false)
│   %9  = Core.apply_type(%7, %8)::Core.Compiler.Const(var"#9#10"{Tuple{Int64,Int64,Int64}}, false)
│         (#9 = %new(%9, P))
│   %11 = #9::var"#9#10"{Tuple{Int64,Int64,Int64}}
│   %12 = Main.Val($(Expr(:static_parameter, 1)))::Core.Compiler.Const(Val{3}(), true)
│   %13 = Main.ntuple(%11, %12)::Tuple{Int64,Int64,Int64}
└──       return %13

```