@mcabbott 2020-03-23
Improvements would be nice to have. On short tuples, but not long ones, [TupleTools.jl](https://github.com/Jutho/TupleTools.jl) is quite a bit quicker:
```julia
julia> @btime fast_invperm((3,1,2,4))
  61.170 ns (2 allocations: 128 bytes)
(2, 3, 1, 4)

julia> @btime TupleTools.invperm((3,1,2,4))
  1.422 ns (0 allocations: 0 bytes)
(2, 3, 1, 4)

julia> @btime fast_invperm((3,1,2,4,5,6,7,8,10,9));
  95.121 ns (2 allocations: 128 bytes)

julia> @btime TupleTools.invperm((3,1,2,4,5,6,7,8,10,9));
  425.181 ns (25 allocations: 1.00 KiB)
```