@mcabbott 2020-03-23
I think that what I wrote is equivalent to `$`. You do get slightly longer times with e.g. `@btime TupleTools.invperm(p)  setup=(p=(3,1,2,4));` or `@btime TupleTools.invperm($(Ref((3,1,2,4)))[]);` as these (if I understand right) circumvent constant propagation. I think we care about the speed both with and without this. But however you slice it, `TupleTools` is beaten from length=10.