@francescoalemanno 2020-03-23
Jutho's TupleTools is much faster than both base and my code for small tuples, since it uses "lispy" tuple programs that rely on dispatch,  this is also the reason that it becomes slower for bigger tuples.
the implementations are quite a lot different. What do you suggest?