@mcabbott 2020-03-23
We should switch, do a lispy thing for small tuples, and something else for large ones. Many functions in Base do this. At length=10, `Base.invperm` is already faster on an Array than on a Tuple, so perhaps we should just call that & convert back.