@mcabbott 2020-03-23
I think it goes faster if the inner for loop is also an ntuple, and you can check invperm inside? There should be no need for `@inbounds` either. Like this, but double-check:
```julia
function fast_invperm3(P::NTuple{N,Int}) where {N}
	ntuple(N) do i
	    s = sum(ntuple(N) do j
	        P[j]==i && return j
	        0
	    end)
	    s == 0 && error("not a perm")
	    s
	end
end
```