@francescoalemanno 2020-03-23
```julia
using Random
using BenchmarkTools

function fast_invperm(P::NTuple{N,T}) where {N,T}
    ntuple(Val(N)) do i
        for j in P
            @inbounds P[j]==i && return j
        end
        throw(ArgumentError("argument is not a permutation"))
    end
end

function time_impl()
    for N in 3:10:50
        T=(shuffle(1:N)...,)
        @show @benchmark invperm($T)
        @show @benchmark fast_invperm($T)
    end
end

time_impl()

#= :100 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(260.141 ns)
#= :101 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(7.581 ns)
#= :100 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(453.706 ns)
#= :101 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(41.056 ns)
#= :100 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(652.377 ns)
#= :101 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(110.756 ns)
#= :100 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(1.303 μs)
#= :101 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(241.758 ns)
#= :100 =# @benchmark(invperm($(Expr(:$, :T)))) = Trial(1.609 μs)
#= :101 =# @benchmark(fast_invperm($(Expr(:$, :T)))) = Trial(415.975 ns)
```