@mcabbott 2020-03-23
One reason to care a lot about very short tuples is things like #33435: PermutedDimsArray{T,4} is common, while PermutedDimsArray{T,14} may never have been spotted in the wild. Getting sub-ns `@btime` is (if I understand right) a sign that constant-propagation is allowing this to be compiled away.