@francescoalemanno 2020-03-23
@mcabbott i agree, i will try to make a compact tuple program to solve it in small cases without overhead from the loop