@francescoalemanno 2020-03-23
@mcabbott, yes i can see that your code for smaller tuples is able to make the compiler do all the heavy lifting, but i do not like the use of sums.