@francescoalemanno 2020-03-24
added `isperm` and also managed to make both method even faster

```julia
function fast_isperm(P::NTuple{N,T}) where {N,T}
    all(ntuple(Val(N)) do i
        for j in eachindex(P)
            @inbounds P[j]==i && return true
        end
        return false
    end)
end

function fast_invperm(P::NTuple{N,T}) where {N,T}
    ntuple(Val(N)) do i
        for j in eachindex(P)
            @inbounds P[j]==i && return T(j)
        end
        throw(ArgumentError("argument is not a permutation"))
    end
end

using Random
using BenchmarkTools


function time_impl()
    for N in 3:10:50
        T=(shuffle(1:N)...,)
        @info "Testing size $N :"
        @btime invperm($T)
        @btime invperm($T)
        @btime fast_invperm($T)
        @btime fast_invperm($T)
    end
end

time_impl()
```

```
[ Info: Testing size 3 :
  249.991 ns (3 allocations: 256 bytes)
  254.721 ns (3 allocations: 256 bytes)
  6.849 ns (0 allocations: 0 bytes)
  6.849 ns (0 allocations: 0 bytes)
[ Info: Testing size 13 :
  458.505 ns (3 allocations: 496 bytes)
  445.848 ns (3 allocations: 496 bytes)
  24.597 ns (0 allocations: 0 bytes)
  24.591 ns (0 allocations: 0 bytes)
[ Info: Testing size 23 :
  641.355 ns (3 allocations: 736 bytes)
  652.488 ns (3 allocations: 736 bytes)
  64.287 ns (0 allocations: 0 bytes)
  64.280 ns (0 allocations: 0 bytes)
[ Info: Testing size 33 :
  1.303 μs (4 allocations: 1.25 KiB)
  1.283 μs (4 allocations: 1.25 KiB)
  222.649 ns (0 allocations: 0 bytes)
  222.863 ns (0 allocations: 0 bytes)
[ Info: Testing size 43 :
  1.611 μs (4 allocations: 1.59 KiB)
  1.614 μs (4 allocations: 1.59 KiB)
  372.524 ns (0 allocations: 0 bytes)
  371.869 ns (0 allocations: 0 bytes)
```